Update engines.md
I was stuck on the "Mounting the Engine" section, first because I didn't understand the Gemfile was the name of the file I needed to change, then because I didn´t know I had to create a new folder called 'engines' and then copy the 'blorh' files there. I think the changes I propose help to clarify that.